pyproject.toml
[tool.poetry]
name = "vlan_sync"
version = "0.1.0"
description = ""
authors = ["adrian <aegiacometti@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
netmiko = "^3.4.0"
PyYAML = "^6.0"
SQLAlchemy = "^1.4.26"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
